requirements.txt
beautifulsoup4==4.13.5
langchain==0.3.27
langextract==1.0.9
langextract==1.0.8
langgraph==0.6.7
motor==3.7.1
nest_asyncio==1.6.0
numpy==2.3.3
pandas==2.3.2
protobuf==6.32.1
pytest==8.3.5
python-dotenv==1.1.1
Requests==2.32.5
scikit_learn==1.7.2
streamlit==1.47.1
streamlit==1.45.1
